Can now move anywhere in 3D space
The ship can now roll clockwise and counter clockwise in addition to
turning (yawing?) left and right. This gives it access to all of
3D space in a somewhat awkward way. Before this, it was confined to
a plane.

This commit also includes fixes and generalizations to the first-person
camera tracking strategy so the camera's own up vector stays fixed to
the ship's.